requirements.txt
fastapi
uvicorn[standard]
transformers
scikit-learn
pandas
mastodon.py
python-multipart
python-dotenv
gdown